+++ CHANGELOG.md
@@ -40,6 +40,7 @@ The releases of Garlmap aim to follow [semantic versioning](https://semver.org).
 - Shortcut for exporting the playlist is now Ctrl-t instead of the weird Ctrl-x that prevented "cut"
 - Fields "disc_total" and "track_total" are now named "disctotal" and "tracktotal"
 - Minimum window size is now 320 instead of 700, because window resizing is improved
+- Rule input field is now resizable and handles pasted newlines correctly
 
 ### Fixed
 
@@ -54,7 +55,7 @@ The releases of Garlmap aim to follow [semantic versioning](https://semver.org).
 
 ### Versions
 
-- Electron 19.0.3 (was 18.1.0)
+- Electron 19.0.4 (was 18.1.0)
 - Chromium 102.0.5005.63 (was 100.0.4896.127)
 
 ## [0.4.0](https://github.com/Jelmerro/Garlmap/compare/0.3.0...0.4.0) - 2022-04-27